feat: add semantic token range endpoint
